apply transforms on double tensors

+++ src/diffoptics/transforms/BaseTransform.py
@@ -6,29 +6,37 @@ from diffoptics.optics.Ray import Rays
 class BaseTransform(abc.ABC):
 
     def apply_transform(self, rays: Rays):
-        new_o = torch.matmul(self.transform.type(rays.origins.dtype).to(rays.device),
-                             torch.cat((rays.origins, torch.ones((rays.origins.shape[0], 1),
-                                        device=rays.device)), dim=1).unsqueeze(-1))[:, :3, 0]
-        new_d = torch.matmul(self.transform.type(rays.directions.dtype).to(rays.device),
-                             torch.cat((rays.directions, torch.zeros((rays.directions.shape[0], 1),
-                                        device=rays.device)), dim=1).unsqueeze(-1))[:, :3, 0]
+        new_o = torch.matmul(self.transform.to(rays.device),
+                             torch.cat((rays.origins.type(torch.double), torch.ones((rays.origins.shape[0], 1),
+                                                                                    device=rays.device)),
+                                       dim=1).unsqueeze(-1))[:, :3, 0].type(rays.origins.dtype)
+        new_d = torch.matmul(self.transform.to(rays.device),
+                             torch.cat((rays.directions.type(torch.double),
+                                        torch.zeros((rays.directions.shape[0], 1), dtype=torch.double,
+                                                    device=rays.device)), dim=1).unsqueeze(-1))[:, :3, 0].type(
+            rays.directions.dtype)
         return Rays(new_o, new_d, luminosities=rays.luminosities, meta=rays.meta, device=rays.device)
 
     def apply_inverse_transform(self, rays: Rays):
-        new_o = torch.matmul(self.inverse_transform.type(rays.origins.dtype).to(rays.device),
-                             torch.cat((rays.origins, torch.ones((rays.origins.shape[0], 1),
-                                        device=rays.device)), dim=1).unsqueeze(-1))[:, :3, 0]
-        new_d = torch.matmul(self.inverse_transform.type(rays.directions.dtype).to(rays.device),
-                             torch.cat((rays.directions, torch.zeros((rays.directions.shape[0], 1),
-                                        device=rays.device)), dim=1).unsqueeze(-1))[:, :3, 0]
+        new_o = torch.matmul(self.inverse_transform.to(rays.device),
+                             torch.cat((rays.origins.type(torch.double),
+                                        torch.ones((rays.origins.shape[0], 1), dtype=torch.double,
+                                                   device=rays.device)), dim=1).unsqueeze(-1))[:, :3, 0].type(
+            rays.origins.dtype)
+        new_d = torch.matmul(self.inverse_transform.to(rays.device),
+                             torch.cat((rays.directions.type(torch.double), torch.zeros((rays.directions.shape[0], 1),
+                                                                                        device=rays.device)),
+                                       dim=1).unsqueeze(-1))[:, :3, 0].type(rays.directions.dtype)
         return Rays(new_o, new_d, luminosities=rays.luminosities, meta=rays.meta, device=rays.device)
 
     def apply_transform_(self, points: torch.tensor):
-        return torch.matmul(self.transform.type(points.dtype).to(points.device),
-                            torch.cat((points, torch.ones((points.shape[0], 1),
-                                                          device=points.device)), dim=1).unsqueeze(-1))[:, :3, 0]
+        return torch.matmul(self.transform.to(points.device),
+                            torch.cat((points.type(torch.double), torch.ones((points.shape[0], 1), dtype=torch.double,
+                                                                             device=points.device)), dim=1).unsqueeze(
+                                -1))[:, :3, 0].type(points.dtype)
 
     def apply_inverse_transform_(self, points: torch.tensor):
-        return torch.matmul(self.inverse_transform.type(points.dtype).to(points.device),
-                            torch.cat((points, torch.ones((points.shape[0], 1),
-                                                          device=points.device)), dim=1).unsqueeze(-1))[:, :3, 0]
+        return torch.matmul(self.inverse_transform.to(points.device),
+                            torch.cat((points.type(torch.double), torch.ones((points.shape[0], 1), dtype=torch.double,
+                                                                             device=points.device)), dim=1).unsqueeze(
+                                -1))[:, :3, 0].type(points.dtype)
